Moves type to separate file.
