+++ plugins/woocommerce/client/blocks/assets/js/blocks/product-gallery/frontend.ts
@@ -12,53 +12,13 @@ import {
 /**
  * Internal dependencies
  */
-import type { ProductGalleryContext, ImageDataItem } from './types';
+import type {
+	ProductGalleryState,
+	ImageDataItem,
+	ProductGalleryStore,
+} from './types';
 
-interface StoreState {
-	imageData: ImageDataItem[] | undefined;
-	selectedImageId: number;
-	isDialogOpen: boolean;
-	productId: string;
-	disableLeft: boolean;
-	disableRight: boolean;
-	touchStartX: number;
-	touchCurrentX: number;
-	isDragging: boolean;
-	readonly allImageIds: number[];
-	thumbnails: () => ImageDataItem[] | undefined;
-}
-
-interface StoreActions {
-	selectImage: ( newImageNumber: number ) => void;
-	selectCurrentImage: ( event?: MouseEvent ) => void;
-	selectNextImage: ( event?: MouseEvent ) => void;
-	selectPreviousImage: ( event?: MouseEvent ) => void;
-	onSelectedLargeImageKeyDown: ( event: KeyboardEvent ) => void;
-	onViewAllImagesKeyDown: ( event: KeyboardEvent ) => void;
-	onThumbnailKeyDown: ( event: KeyboardEvent ) => void;
-	onDialogKeyDown: ( event: KeyboardEvent ) => void;
-	openDialog: () => void;
-	closeDialog: () => void;
-	onTouchStart: ( event: TouchEvent ) => void;
-	onTouchMove: ( event: TouchEvent ) => void;
-	onTouchEnd: () => void;
-	displayViewAll: () => boolean;
-}
-
-interface StoreCallbacks {
-	watchForChangesOnAddToCartForm: () => void;
-	dialogStateChange: () => void;
-	toggleActiveImageAtrributes: () => void;
-}
-
-interface ProductGalleryStore {
-	state: StoreState;
-	actions: StoreActions;
-	callbacks: StoreCallbacks;
-}
-
-const getContext = ( ns?: string ) =>
-	getContextFn< ProductGalleryContext >( ns );
+const getContext = ( ns?: string ) => getContextFn< ProductGalleryState >( ns );
 
 const getArrowsState = ( imageId: number ) => ( {
 	disableLeft: imageId === state.allImageIds[ 0 ],
