+++ plugins/woocommerce/client/blocks/assets/js/blocks/product-gallery/types.ts
@@ -18,11 +18,8 @@ export interface ImageDataItem {
 	sizes: string;
 }
 
-interface ImageData {
-	images: ImageDataItem[];
-}
-
-export interface ProductGalleryContext {
+export interface ProductGalleryState {
+	imageData: ImageDataItem[] | undefined;
 	selectedImageId: number;
 	isDialogOpen: boolean;
 	productId: string;
@@ -31,6 +28,35 @@ export interface ProductGalleryContext {
 	touchStartX: number;
 	touchCurrentX: number;
 	isDragging: boolean;
-	imageData: ImageData;
-	image: ImageDataItem;
+	readonly allImageIds: number[];
+	thumbnails: () => ImageDataItem[] | undefined;
+}
+
+interface StoreActions {
+	selectImage: ( newImageNumber: number ) => void;
+	selectCurrentImage: ( event?: MouseEvent ) => void;
+	selectNextImage: ( event?: MouseEvent ) => void;
+	selectPreviousImage: ( event?: MouseEvent ) => void;
+	onSelectedLargeImageKeyDown: ( event: KeyboardEvent ) => void;
+	onViewAllImagesKeyDown: ( event: KeyboardEvent ) => void;
+	onThumbnailKeyDown: ( event: KeyboardEvent ) => void;
+	onDialogKeyDown: ( event: KeyboardEvent ) => void;
+	openDialog: () => void;
+	closeDialog: () => void;
+	onTouchStart: ( event: TouchEvent ) => void;
+	onTouchMove: ( event: TouchEvent ) => void;
+	onTouchEnd: () => void;
+	displayViewAll: () => boolean;
+}
+
+interface StoreCallbacks {
+	watchForChangesOnAddToCartForm: () => void;
+	dialogStateChange: () => void;
+	toggleActiveImageAtrributes: () => void;
+}
+
+export interface ProductGalleryStore {
+	state: ProductGalleryState;
+	actions: StoreActions;
+	callbacks: StoreCallbacks;
 }
